@@ -5,9 +5,8 @@ namespace Dashifen\Router\Route;
 use Dashifen\Exception\Exception;
 
 class RouteException extends Exception {
-  public const UNEXPECTED_METHOD = 1;
-  public const INVALID_PATH = 2;
-  public const INVALID_ACTION = 3;
-  public const UNKNOWN_PROPERTY = 4;
-  public const INVALID_CONSTRUCTION = 5;
+  public const int UNKNOWN_METHOD       = 1;
+  public const int UNKNOWN_PATH         = 2;
+  public const int UNKNOWN_ACTION       = 3;
+  public const int UNKNOWN_PROPERTY     = 4;
 }
