@@ -11,21 +11,15 @@ use Dashifen\Router\Route\RouteInterface;
  */
 interface RouterInterface {
   /**
-   * isAutoRouter
-   *
-   * An "auto-router" automatically constructs the RouteInterface based
-   * on the current request.  There's no need to add routes to its collection,
-   * it'll figure it all out on its own!
-   *
-   * @param bool|null $autoRouterState
+   * Returns true if this is an "auto-router."  An auto-router determines the
+   * route all on its own based on the current request; you don't have to
+   * hand-enter routes for it.
    *
    * @return bool
    */
-  public function isAutoRouter(?bool $autoRouterState = null): bool;
+  public function isAutoRouter(): bool;
 
   /**
-   * getRoute
-   *
    * Returns the route for the current request; will construct that route
    * if this is an auto-router.
    *
@@ -34,8 +28,6 @@ interface RouterInterface {
   public function getRoute (): RouteInterface;
 
   /**
-   * getRouters
-   *
    * Returns the full set of routes in the router's collection.  Throws an
    * exception if this is an auto-router.
    *
@@ -45,8 +37,6 @@ interface RouterInterface {
   public function getRoutes (): array;
 
   /**
-   * addRoute
-   *
    * Adds a route to this router's collection based on the indices of the
    * parameter array, but throws an exception fi this is an auto-router.
    *
@@ -57,8 +47,6 @@ interface RouterInterface {
   public function addRoute (array $route): void;
 
   /**
-   * addRoutes
-   *
    * Adds the specified routes to this router's collection, but throws an
    * exception if this is an auto-router.
    *
